HACK: plymouth: force scale to 1 when there aren't system-updates scheduled
plymouth doesn't support HiDPI resources, meaning that images (such as the Droidian
logo) would appear blurry due to upscaling (even when not needed!).

Forcibly set the device scale to 1, both on initramfs (this can be overridden via the
standard plymouth.force-scale cmdline argument) and on systemd.

When a system-update is scheduled, quit plymouth on initramfs and let systemd
decide whether to force scaling or not.

This is an hack, and breaks other themes (that would be still broken due to color channels
being swapped), and adds a flicker when updating. The upside is that in the majority
of cases, Droidian logo is crisp (and we avoid doing unnecessary scaling shenanigans).

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>